Move String.load from JsUnpacker to AppContextUtils
It is only used there, is not really a fir in library, and also I am very confused what this is even supposed to do.
